requirements.txt
python-dotenv~=1.0.1
pymongo~=4.7.2
flask~=3.0.3